[rev] delete old `ImageObliqueSliceViewModel`
